Updating frontend compile signature (#30)
This change updates the signature of `compile` in the qsc_frontenc to take an array of inputs and an optional expression input for execution. This is a steppingstone for implementing the interpreter in qsc_eval, but still leaves several elements for local compilation to be done in follow up PRs, in particular tracking `Span` for AST nodes on a per-input level instead of in the aggregate to make it easy to correlate back to individual files.