Initial Evaluator implementation (#31)
This change adds the initial implementation for the evaluator. This does
not yet include any state tracking and just introduces handling for
literals, array creation and indexing, tuples, ranges, and fail
expressions. It demonstrates a pattern for execution, testing and error
reporting that can be expanded upon in later PRs.

---------

Co-authored-by: Sarah Marshall <33814365+samarsha@users.noreply.github.com>
Co-authored-by: Sarah Marshall <sarah.marshall@microsoft.com>