Parse array repeat and lambda expressions (#39)
Lambdas are parsed as an infix operator, so the LHS is an expression.
This needs to be converted into a pattern when the lambda arrow is
encountered. The benefit of this is that it doesn't require
backtracking, but the downside is that the grammars for expressions and
patterns don't necessarily align. It happens to work OK for now, though,
based on the patterns for lambdas that the old Q# compiler supports.