Improve global name resolution (#60)
This improves the shadowing rules when resolving globals to avoid
ambiguous name errors when there is an intuitive choice. Generally, the
rule is still that nearer definitions shadow farther definitions. This
is true for local scopes, and I extended it to globals:

- Items in the parent namespace shadow items in opened namespaces.
- Items in an aliased namespace shadow unopened globals that happen to
share the same path.

The only case where a name is ambiguous is when there is a choice
between opened namespaces.