Support for call expression, Adjoint functor (#68)
This change introduces support for calling operations and functions,
including argument and return support, as well as support for tracking
`Adjoint` and `Controlled` functors on callable values. It assumes the
presence of a specialization implementation, except in the case of the
`intrinsic` generator which expects to match the call by name with one
natively supported in the evaluator. Adjoint calls are supported, but
Controlled calls is unimplemented pending follow up with support for
nested of argument tuples in the invocation.

---------

Co-authored-by: Sarah Marshall <33814365+samarsha@users.noreply.github.com>