Add Single Qubit Gates Katas, basic validation
This change adds the tasks, reference implementation, and basic validation for the Single Qubit Gates (tutorial) kata. The validation is basic in the sense that it only uses validation of the adjoint without exercising the controlled specialization so global phase is not checked. Likewise, tasks 6 and 7 perform validtion only on a single set of inputs rather than looping over a set of possible inputs. It avoids the need for specialization generation by explicitly defining the adjoint of the reference implementation using gates with known support, and each validate operation avoids conjugate expressions and uses nested if-expressions rather than Boolean-and.

This change also updates the standard library to add trigonometric intrinsic functions to Microsoft.Quantum.Math.