npm package and playground (#82)
This PR adds the start of an npm package which includes the qsharp wasm
package and wrapper JS code for Node.js and the browser, as well as the
start of a simple playground site for ad-hoc validation and an example
of building a web site using the npm package.

Opening as a draft for discussion. Still to do a bit of clean-up and add
tests to both projects.