Support Quantum Gate Intrinsics, Equality + AndL BinOp (#78)
This change supports the quantum gate intrinsics as defined in the
library. For now, it uses the state dump that prints to standard out
until we can agree on the proper API for Rust-based state dump.

In addition, this includes support and tests for the equality and
logical-and binary operators along with tests.