Support controlled functor in calls
This updates the handling of (potentially nested) tuples of controlled arrays and original parameters to calls such that all nested arrays of controls are concatenated and sent to the underlying operation.

This change also updates the evaluator unimplemented error to specifically call out which language feature is not implemented. It also adds some additional validation pass checks in frontened to disallow illegal patterns in specialization declarations.