Evaluator refactor (#83)
This change refactors the evaluator to be stateless, making it easier to
do incremental evaluations in the future. As a result, the `Evaluator`
struct no longer needs to be public and the crate instead exposes a top
level `evaluate` function that mirrors the top level `compile` function
exposed from qsc_frontend. Callers can then decide how they invoke
`evaluate` and what context to provide, which is a building block for
possible scenarios like a REPL.

---------

Co-authored-by: Sarah Marshall <sarah.marshall@microsoft.com>