Add `Receiver` trait for handling output
This change adds a new `Receiver` trait that can be used for accepting output from the evaluator execution, specifically things like `Message` and `DumpMachine` for quantum state. It keeps the types specific to the underlying simulator, such that the it is up to the receiver implementation how to render it for their specific environment.

Also included is the addition of `Message` to the library and new `ToString` function to construct strings for display.

This also makes it easier to write tests that check output that would go to the user besides the final return value from execution.