Add language support document (#121)
This adds a language support doc to make it easier to point folks across
the team at what subset of the language is currently supported vs not.
It's not a formal commitment or planning doc, just an easy way to
communicate. Once all features have moved from "Unsupported" to
"Supported" we can delete the doc.