Support `main` with no args as default entry point
This change adds a new evaluator function, `eval_main` that looks for a single global callable named `main` as the place to start evaluation. It further updates the validation code to enforce a single global callable with the name `main` that has no arguments and a single block body declaration. This allows callers to handle the special case where no entry expression is provided or supported and rely on the existence of a defined `main`.

This fixes #109.