Use once_cell to cache stdlib compilation in wasm (#124)
This adds simple caching of the stdlib compilation during check calls
using OnceCell. Since the stdlib is static in the wasm layer and the
store does not need to be mutable (the compilation unit for the checked
user's code is not added to it), this can be done once for the lifetime
of the loaded package and reused over and over. In local testing, this
produced a 95% perf improvement on the check calls performed as the user
types (from 2.3 ms to 0.1 ms).