Add `EntryPoint` attribute handling
This change adds a pass utility that finds the single entry point in a package or fails if a valid one cannot be uniquely identified. In qsc, if no entry expression is provided, the package is checked for an entry point callable. In WASM, the entry point is required but only used if the entry expression is the empty string.