Use qubit count in state dump (#141)
This adapts the state dump display to use the total qubit count returned
from the simulator to adjust the size of the displayed binary, making it
easier to correlate to the qubits allocated at that point in the
program. Depends on updates to the qir_backend.

Includes minor fix to qsi for using normal display instead of debug
print for values.