Clean-up todos, add `newtype` validation errors
This change adds URLs for matching issues to a number of TODO comments in the code. In addition, it includes `newtype` in the unsupported features flagged by validation. It also removes the `HasFunctorsIfOp` type class and related infrastructure since functor support for calls in generated specializations will be validated by specialization generation instead of type checking.