Parse some expression kinds as statement-final (#162)
This PR classifies a few expression kinds (that syntactically always end
with a block) as "statement-final," meaning that if they occur at the
top level of an expression statement, they terminate parsing of the
expression - no operators can appear after them. Those expressions are
`Block`, `Conjugate`, `For`, `If`, `Repeat` and `While`. This fixes
#139.

This PR also adds a check that all but the last statement in a block
must end in a semicolon, *unless* it's a statement-final expression
statement, in which case the semicolon is optional. This allows:

```qsharp
{
    if x { f() }
    if y { g() }
}
```

but forbids:

```qsharp
{
    f()
    g()
}
```