Update error patterns
This change updates the error patterns in the compiler to favor only using labels with strings that add more information to the top level error message and make sense when concatenated. It also updates the `qsc_wasm` error translation to use the label message and help text if present.

Note this updates the `Ambiguous` error to use a single span by referring to other namespaces by name rather than position.