Reduce number of distinct type variables from diverging terms (#196)
The old algorithm for handling divergence during type inference was
technically correct, but created too many new, unconstrained type
variables as a diverging term propagates up the tree, when it would be
possible to reuse the same type. This could cause issues for some passes
that don't expect to see an inference variable in a type.

This PR reduces the number of these unconstrained type variables that
are created. The only ones left should be when the type is truly
unconstrained, such as when the value of the expression is discarded by
a semicolon statement.

It also simplifies the code for divergence analysis and makes it (IMO)
more readable.