Simplify compiler-generated IDs for global declarations (#200)
This PR creates a new ID type for global declarations that is
independent of the declaration's assigned node ID in AST or HIR. This
new ID is assigned by the name resolver, and is baked into HIR, where
it's used as an index in an IndexMap of all items in the package. This
has some advantages:

- Being independent of node IDs, it's easier to persist across IRs. This
will make adding UDTs to the type checker's `Ty` type easier.
- Being an index into an IndexMap, it allows fast lookup of any global
in any package in the package store given only its ID. This makes
`qsc_passes::globals::extract_callables` obsolete.

This PR also refactors the way the evaluator looks up globals, allowing
it to own the incrementally compiled `CallableDecl`s, which allows
removing a usage of `Box::leak`.