Support Adjoint invert spec gen via logic separation analysis (#190)
This change introduces logic separation as an analysis pass that returns
the set of stmts within a block and all nested scopes that are
non-deterministic due to including operation calls. It uses this
analysis pass as the chance to collect any errors related to patterns
that prevent separation and therefore adjoint generation. This is used
by `spec_gen` to support adjoint invert and controlled adjoint invert
generation. Includes tests for logic separation analysis and adjoint
generation. This also includes loop reversal of for-loops over ranges
and arrays.

---------

Co-authored-by: Sarah Marshall <33814365+samarsha@users.noreply.github.com>