Refactor npm package and add Worker support (#203)
This PR cleans up the npm package so nearly all code is shared between
node.js and the browser, with just a few lines in platform specific
entry points. The platform specific entry points are now handled with
'conditional exports' in the package.json, so either Node.js or browser
source consuming the package should just "import { ... } from 'qsharp'".

This PR also adds support for Worker threads, and moves the playground
to use WebWorkers by default (getting all the wasm work such as
compiling and simulating off the main UX thread). As this works
similarly in both browser and Node.js, you can see in some of the unit
tests added under ./npm/tests how this is used. (It's mostly the same as
non-worker usage, just with one-line difference in initialization and
clean-up.

As part of getting the above working, I also added configurable logging
support. See ./npm/src/log.ts for details on that. (It should be pretty
easy to wire this up to both VS Code "output channels" and macros from
the Rust 'log' crate also).

I also made a lot of the TypeScript code more 'strongly typed', and it
deals with a lot of string parsing and message handling in the 'npm'
package now, avoid the need for consumers to do so.

Running the katas works, but I need to clean up the integration in the
playground still for some of the refactoring before they are showing in
the playground again (just commented out for now). I'll do that as part
of a bigger playground UX clean-up.

I touched nearly every line of code in the npm package, so probably just
easier to look at the code overall there, rather than the diff. Sorry
:-/