Lower fields in HIR and type check open ranges (#204)
Add an enum to HIR for the built-in type fields `Length`, `Start`,
`Step` and `End` to avoid stringly-typed fields.

Trying to access the `Start` or `End` fields for ranges that don't have
them is now a type error. This is done by adding three new range types:
RangeTo (`...n`), RangeFrom (`n...`), and RangeFull (`...`). These are
special compiler-internal "Types Who Must Not Be Named." This
effectively limits the use of open-ended ranges to as an array slice,
because the only thing they can do is satisfy the `Index` class.

The old compiler limited the usage of open-ended ranges by making it a
parse error if they occur outside an index, and then by automatically
inserting generated `0` and/or `Length(array)` expressions into the
range slots to make it appear like a normal range. IMO, I think solving
it through types is cleaner, but does come at the cost of 3 new types.