Items as statements (#252)
Extend the grammar to support items (`open`, `newtype`, `function` and
`operation`) as statements in a block.

This makes Q# more expressive since it's convenient to be able to
declare local helper functions in a limited scope like you can in many
other languages.

It also makes the interpreter more usable while simplifying its
implementation. It brings the REPL and notebooks closer to feature
parity with file-based Q#. You can now freely mix multiple callable
declarations and block statements in a single REPL line or notebook
cell. Opening namespaces is also now supported.

With this change, I think the only thing that is keeping a notebook cell
from being as powerful as a file is the ability to declare namespaces. I
think this won't be too hard to add later.

This is a low risk language change. It's backwards compatible and
doesn't introduce any new semantics; it's just syntactic sugar for
global declarations that you can't access from outside a specific scope.
Because lowering to HIR already flattens nested items, the impact is
limited to name resolution and type checking, with passes and the
evaluator being completely unaffected.