Add ESLint and fix issues (#272)
If you have the ESLint extension installed (or run it from the
command-line) it will now give the usual warnings/errors.

It's clean as of this PR, but I'm not making it part of any build or CI
yet. We can determine if we want to add that check after.

I'll work on Prettier/formatting after this.