Implement `ApplyToEach` and `MeasureEachZ` (#287)
This change implements the `ApplyToEach` and `MeasureEachZ` operations.

The `ApplyToEach` implementation differs from the original
implementation to avoid create specific variants depending on the
functors of the operation to apply (suffixed A, C & CA). Instead, it
relies on the runtime to catch missing specialization errors. These
cases are added as tests to convey the new expected behavior. A more
robust solution would be something like functor generics.

---------

Co-authored-by: Cesar Zaragoza Cortes <cesarzc@ntdev.microsoft.com>