Remove `Length` field (#308)
`Length` is not like the other fields. While `array::Length` makes
sense, `array w/ Length <- x` doesn't - you can't "update" the length of
an array to be a different value. You can resize the array with slicing
or concatenation, but that's not like any other kind of field update,
and growing an array requires a fill value, which can't be provided with
the `w/` operator.

Allowing `::Length` but forbidding `w/ Length` creates an inconsistency
in the language. All fields are gettable and updatable - except for
`Length`. That's an indication that either a field isn't the right
abstraction for `Length`, or that fields in Q# are more complicated than
we thought (for example, supporting independent getters and setters a la
C# properties). Either way, it's a language design issue that we haven't
fully thought about. We should be cautious and back this out for now.