Simplify attributes, visibility and functor expressions in HIR (#300)
This takes care of three low-hanging fruit tasks from #295:

* Lower attributes into an enum containing all known attributes and
their arguments, since each instance of an attribute is known at compile
time.
* Lower visibility by removing the wrapper struct and the `Option`,
defaulting to public.
* Lower functor expressions by evaluating them into a set. Also, add
functors implied from specializations, so that the `functors` field in
`CallableDecl` contains *all* of the functors the callable supports.

Because lowering functor expressions removes spans, it would be harder
for the semantic validation pass to generate errors. But I think it's
wrong for that pass to work on HIR. I moved it from qsc_passes to
qsc_frontend and changed it to an AST visitor. I also renamed it from
`semantics` to `funop` (functions and operations), because `semantics`
is too vague. It's specifically checking restrictions for functions and
operations.