User-defined types (#304)
Full support for user-defined types. Closes #148.

UDTs are erased at runtime and aggressively lowered in HIR into
something that is basically a tuple, with field names becoming numbered
indices. This means that only a small change is needed in the evaluator.

The heuristic for disambiguating the ternary update operator into an
index update or a field update is the same as the old compiler.