Compress code link (#318)
With the Shor's algorithm sample, the "share code" link was so big it
didn't work anymore. This change uses gzip compression on the data
before encoding it. With this change, the code link for Shor's went from
30,990 chars to 8,280 chars, so we should be good on code size for a
while.

Note this is only recently supported in 'modern' browsers, but should
work in any recent Chrome, Edge, or Safari.

I also added the popup on copying the link to the clipboard.