Add git hash to version info
This change adds the git hash to the version info in the cmdline tools, as well as making it available from the developer console in the web via a function call on the logger. The logger function approach was a path of least resistance choice, so I'm happy to pivot to another approach if that would be preferable!