Add tasks for full build and playground dev (#329)
Opening this partly to start discussion on if we want these shared tasks
checked in.

This adds a "Full build" task you can run from VS Code (that does the
same as running ./build.py from the command line), and also a "Watch and
serve playground" combined task I often use while developing others may
find uses (it watches/rebuilds npm and playground in the background, so
you just save edits and refresh the browser while working).

I didn't make anything a 'default' build task, to avoid conflicting with
anyone's local tasks. (And as this is in the 'shared' tasks.json, it
will have no effect if not using the "Workspace Config+" extension, and
will combine with your tasks.local.json if you are).