Move `funop` to a pass, fix pass errors in qsi
This change updates the `funop` analysis previously done in the frontend to become a pass performed on the HIR. In the process, it also updates the error reporting from incremental compilation to correctly return lowering errors and include pass errors from the stateful interpreter.

Note that this change also does away with the limitation on using `while` loops from operations, such that operations (and the global scope used by the interpreter) are a superset of what can be used inside of a function.