Update type inference ordering for clearer errors
This change updates the order in which type inference constraints are solved, helping create clearer error messages. This also includes an update to how block types are inferred such that blocks that have a known type are treated as that type, making any type mismatch occur within the block instead of on the whole block.

Fixes #309