Fix instantiation of type schemes with duplicate parameter names (#382)
The root cause here is that callables with duplicate type parameter
names are allowed to be declared (#195). Fixing that is a bigger change.
In the meantime, I don't think it's fair for `Scheme::instantiate` to
complain about an arity mismatch when the caller did actually provide
the right number of arguments (it's only the constructed hash map that
is smaller). Check the arity based on the original input list instead of
the hash map.

Fixes #381.