Error squiggles in VS Code (#394)
This PR accomplishes two main things:

1. Use ESBuild to bundle the extension JavaScript into a single file,
and load the Q# WASM module in the VS Code extension that's _hopefully_
compatible with VS Code Web. The web behavior is still untested as of
this PR. The extension does work with local VS Code.

2. Use `checkCode()` (same method used by the playground) to surface
diagnostics in VS Code.
- Bonus: update the `severity` field in `VSDiagnostic` to be a string
enum since
[everyone](https://docs.rs/miette/latest/miette/enum.Severity.html#)
[seems](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#diagnosticSeverity)
[to](https://github.com/microsoft/vscode/blob/main/src/vscode-dts/vscode.d.ts#L6068)
[disagree](https://microsoft.github.io/monaco-editor/typedoc/enums/MarkerSeverity.html)
on what the enum values should be.