JS bindings for language service (#401)
This adds a wrapper over `LanguageService` in the `wasm` crate to expose
it as a JavaScript class. See: #396

The existing WASM methods `check_code()` and `get_completions()` are
made redundant now that we have the language service, so I'm removing
them and adding a shim where they're used in the npm package.

Exactly how we're going to expose the language service in the npm
package is still being discussed, so I thought I'd get the more
straightforward WASM changes checked in first.