Document LL(1) Scanner behavior (#425)
This PR documents the LL(1) behavior of scanner. It also somewhat tests
that there is no .clone()
method on Scanner.

~I somewhat expect that this PR will receive feedback that modifying the
API just to test that something isn't
.clone() is not worth it. I somewhat agree with that. But, this is kind
of a cool little test. Maybe it is worth
keeping. At the least, I think we should keep the docstring.~

This PR now just has the docstring.