Introduce `Backend` trait for eval
This change updates the way the evaluator targets the quantum backend by introducing a new trait, `Backend`, that is provides the quantum functionality for the evaluation. The sparse simulator becomes one option for a backend, and other implementers could be included in the future.

As written, this is dependent on https://github.com/qir-alliance/qir-runner/pull/84, since it avoids a dependency on either the qir-backend or qir-stdlib crates in favor of building against just the sparse simulator alone.