Katas composition using a single markdown file (#434)
This change refactors the way katas composition is done.

Before, composition of katas was done by creating folders that contained
a specific set of files to represent a kata section. The order of
sections was determined by a JSON file that listed the folders that
composed the kata.

Now, composition of katas is done through a single markdown file. The
markdown file that represents a kata contains macros such as
`@[Example]` and `@[Exercise]` which represent special sections that
should be shown in that particular segment of the kata.

Additionally, the npm package APIs were improved.

---------

Co-authored-by: Cesar Zaragoza Cortes <cesarzc@ntdev.microsoft.com>
Co-authored-by: DmitryVasilevsky <60718360+DmitryVasilevsky@users.noreply.github.com>