Telemetry (#397)
Initial outline for feedback. 

I put this in the most depended on crate so any other crate can log
telemetry events. Happy to move this based on feedback to a more
suitable place though.

I can add more comments to explain why due to the JS/Wasm hook-up this
needs to be a thread_local if need be, which makes it a little more
complex that ideal (but it more correct).

The node tests and the playground changes show roughly how this could be
wired up.

---------

Co-authored-by: Mine Starks <mineyalc@microsoft.com>