Don't panic if generic param name isn't in scope
We handled this correctly in the rules, but not in the actual lookup. Basically, we were overly permissive in our
`unreachable` statement -- if a generic param is not in scope, then its lookup results in a `None`, and that is
actually entirely reachable. Returning a `Ty::Err` instead gives us a clean missing type error.

Closes #454