Add QIR base profile validation pass
This change adds an analysis pass that will check whether a given program conforms to the QIR base profile. Specifically, this disallows:

 - Comparison of `Result` values
 - `Result` literals
 - Operation intrinsics that don't return Result or Unit
 - Entry point callables or expressions that return types other than `Result` or aggregates of `Result`

The logic for the pass with validate the entire package provided, but then only validate callables from dependent packages that are called somewhere in the given package. This allows for use of a subset of the standard library without producing extraneous errors.

The new pass is not run by default, but will be triggered if the compiler is invoked with qir emission, ie: `qsc <file> --emit qir`.