Fix entry point error checking (#474)
This change updates the way entry point attributes and/or expressions
are checked to ensure that the error messages get propageted as
expected. Recent changes regressed the error cases, and this introduces
a new parameter to compile that lets the caller decide whether an entry
point attribute or expression should be treated as required. This is
essentially the difference between a library compilation (no entry
required) and and executable compilation (single entry required), so
this code is a placeholder until we formalize the difference between
library compilation and exe compilation.