Fix katas using package type setting (#517)
This is a workaround to fix kata behavior in the playground, where they
were incorrectly treated as exe documents and failed compilation due to
missing entry point. The katas need to be treated as if they are lib
document. This change adds a new parameter to `updateDocument` on the
language server so that the caller can indicate what type the document
is.

Fixes #516

---------

Co-authored-by: Bill Ticehurst <billti@microsoft.com>