Ignore free variables with no local mapping
In rare circumstances, error recovery in the compiler can lead to AST that includes resolved names for otherwise disallowed patterns in specialization declarations. If these bindings are then used in a lamda, lifting will fail to find the corresponding local because invalid patterns are not lowered into HIR and included in locals. This updates the lifting logic to ignore any free variables that don't have a corresponding local binding and leave them as unmapped nodes. This is safe because only already invalid AST can lead to this scenario.

Fixes #524