Fix Python wheel being overwritten by local install (#539)
The editable install (`pip install -e .`) step was somehow overwriting
the `.whl` file that was built in the previous step. I changed it to do
a wheel install instead of a local install.

Also removed the JupyterLab extension integration test since it can fail
in some environments, due to the specific way `build.py` sets up the
virtual environment. The JupyterLab application directory is based on
`sys.prefix` which can change depending on where/how the script is run.
I could not find a reliable way to get the test working, so I decided to
take it out entirely.

In the future, it would be nice to have an integration test run that
sets up a proper environment and does more thorough validation. Until
then, I think we can live without this one test, which was of dubious
value anyway.