Add QIR base profile codegen (#500)
This change adds a simple codegen for QIR base profile. It assumes the
given HIR is valid for QIR base profile checks (see #494), so is able to
use a custom "gate collecting" simulator to record the gate sequence and
relevant measurements, producing the .ll file matching QIR base profile.
This includes tests for content but does not yet include scenario tests
that use LLVM and/or PyQIR to validate the produced module.