Add QIR base profile validation pass (#494)
This change adds an analysis pass that will check whether a given
program conforms to the QIR base profile. Specifically, this disallows:

 - Comparison of `Result` values
 - `Result` literals
 - Operation intrinsics that don't return Result or Unit
- Entry point callables or expressions that return types other than
`Result` or aggregates of `Result`

The new pass is only run if the current target is `Base` as indicated by
arguments to compile. Code can be conditionally compiled out by
including `@Config(Full)` or `@Config(Base)`. This PR adds the logic for
simple conditional compilation and uses it to skip compiling of
callables in the standard library that are not base profile compliant.