Unify file access in debugger (#584)
This is a workaround for windows which doesn't play well with URIs. The
DAP
`setDebuggerPathFormat`/`convertClientPathToDebugger`/`convertDebuggerPathToClient`
can't be used as there is a [bug requiring
nodejs](https://github.com/microsoft/vscode-debugadapter-node/issues/298)
which prevents usage in web extensions. Once the bug is fixed we can
update the debugger to use URIs without this PR's retry conversion.

This PR also fixes an issue where if the user closed the file being
debugged or it was open as 'temporarily open' then we couldn't step back
into the code.