Better BP and stepping through behavior (#582)
Compilation may generate code that we don't want to step into. This PR
up changes the evaluator to continue executing while the current
statement does not have a source span mapping. Additionally, the HIR in
some passes is updated to either set or default the statement spans
depending on whether we want that code to be steppable.

While loops' condition is an expression thus we can't stop on looping at
the conditional check without adding expr stepping or creating a pass
for while loops that can introduce a NOP statement with a span on the
expr.

```
Teleport evaluation     time:   [102.15 µs 102.21 µs 102.29 µs]
                        change: [+0.1592% +0.3292% +0.5028%] (p = 0.00 < 0.05)
                        Change within noise threshold.

Deutsch-Jozsa evaluation
                        time:   [22.519 ms 22.531 ms 22.547 ms]
                        change: [-0.8468% -0.7858% -0.7114%] (p = 0.00 < 0.05)
                        Change within noise threshold.
```

---------

Co-authored-by: Stefan J. Wernli <swernli@microsoft.com>