Python `compile` for QIR codegen (#588)
This adds a `compile` method to the Python `qsharp` module and
introduces an `init` pattern to be used when starting up a session. The
expected usage is that for full compilation and simulation,
`qsharp.init()` or `qsharp.init(target=qsharp.Target.Full)` are called
at the outset to configure compilation (NOTE: latest allows the user to
skip this and just takes the default on first use). If
`qsharp.init(target=qsharp.Target.Base)` is used instead then local
simulation will follow the same restrictions as base profile code
generation. When in this mode, `qsharp.compile(<expr>)` will produce
output that can be used as part of submitting to QIR capable targets.

---------

Co-authored-by: Mine Starks <mineyalc@microsoft.com>