Use `strip = true` to reduce Linux binary sizes
This updates the workspace settings for cargo to remove `debug =1` and instead use `strip = true` to reduce binary sizes. Previously, the Linux .so that was part of the Python wheel was ~31 MB while other platforms were in the single digits. After this change, the same .so is only ~3.2 MB. Perf testing showed no change in behavior, so the only trade-off for the binary size is no longer including debug symbols for the release binaries.