Items shouldn't need to be defined before use in `%%qsharp` cells
This change updates the way fragments are compiled to ensure that each stage of compilation (ie: parsing, item discovery, name resolution, type checking, etc) are performed on all fragments before progressing to the next stage. This allows for the expected Q# patterns to work as expected in cells, namely that namespaces can be mutually recursive and items can be used before they are defined (as long as they are defined later in the same cell). This also adds some additional test cases for non-incremental compilation to verify these behaviors there too.
Fixes #594