Large numbers aren't treated properly (#635)
This change updates the logic for how the parser handles integer
literals to disallow most wrapping. The one corner case that is still
allowed is that `9_223_372_036_854_775_808` or `i64::MAX + 1` wraps to
`i64::MIN`. This is necessary because the negative sign in Q# is parsed
as the unary negate operator, so the literal
`-9_223_372_036_854_775_808` is really a negate of the positive literal.
Using wrapping add, the value becomes `i64::MIN` and then negate of
`i64::MIN` is `i64::MIN` so the right value is achieved at runtime. Any
other values that are too large will be flagged as invalid integer
literals. Existing tests have been updated and new tests added to
confirm the logic for handling min, max, and overflowing values.
Fixes #632