Fix reporting of interpreter errors with spans across multiple lines. (#622)
Fixes most of #596. The one case it doesn't address is when runtime
errors have spans that should point to a different package than the
current one, e.g. stdlib. I will tackle that in a different PR.

Also fixes #595 , minus the parsing of doc comments.

Lots of file changes (sorry, mostly tests!), but there are really only
two key points:
- We now preserve already-executed interpreter lines in a `SourceMap`
object that we keep appending to.
- The `WithSource` error wrapper can now contain more than one source
file, which is handy when the error happens to contain multiple spans
that come from different files.