Update to Rust 1.72.0 (#644)
This update comes with changes to clippy and rustfmt that can cause
validation failures, so the change includes all the updates to source
code and configuration to bring the codebase into alignment with the
latest Rust lints. Without this, new users who install the latest stable
rust would see these failures locally and be blocked from developing.

The prerequisites check is also updated to force everyone to upgrade.