Update Rust dependencies for component governance
This updates the rust dependencies for criterion and pyo3 to the latest versions. Notably, newer versions of criterion do not depend on atty, which was the source of a component governance warning. Also, the dependency on the indenter crate was left as-is since newer versions change the API in a way that will require some refactoring and is better left for a follow up PR.