Add ability to set target profile in language service (#657)
Introducing a `WorkspaceConfiguration` concept to the language service
which provides a central means of setting compiler options, including
`TargetProfile` as well as `PackageType`, which we were previously
passing in `update_document`.

In a future PR, these settings can be wired up to VS Code settings or to
a VS Code command. (Draft PR #663)

I'm keeping compatibility with LSP in mind. The new
`update_configuration` method carries the semantics of the
[`workspace/didChangeConfiguration`](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#workspace_didChangeConfiguration)
notification in LSP.

We don't currently support per-document settings. I'm not sure if we
would want to, given our current user scenarios (where would we persist
such a setting in VS Code?). As we flesh out Q# "projects" and other
ways of passing compiler settings, we can work those into the language
service as well.