Fix infinite type inference recursion (#669)
This change adds a limit on recursive type inference that avoids
situations where solving for an inferred type could lead to infinite
recursion. The test case added, when present in a .qs file in VSCode or
the web, would cause runaway memory usage (observed WASM in the broswer
using the full 2 GB allowed) and slowing down editor performance. This
is similar to the recursive type inference limit used in other languages
like Rust.
Fixes #668