VS Code editor codegen and Azure submission (#621)
A work in progress, but with this you can:

- Run the "Get QIR for current Q# program" to see the LL codegen for the
currently active editor - if base profile compliant. (Note: May not
leave this in, but it's handy for testing and those who are
QIR-curious).
- Go into settings and enable the Azure Integration (need to opt in
currently due to experimental nature and use of temp service for
processing)
- Sign-in to Azure via you AAD or MSA account and see your workspaces
(ignore the 'Connection String' option for now which is mock data for
quick validation)
- Click on the submit icon (looks like a Play button) to the right of a
hardware target to submit the currently active editor code.
 - Refresh the workspace view to see latest job status.
- Click on the download icon to see the raw histogram data for completed
jobs.

It's spaghetti code currently from much experimentation, but sharing for
early feedback on design and for a peek into service integration.

Still TODO for this PR:

- [x] Wire up the QIR Profile selector to switch the language service
between base and full profiles and show corresponding errors.
- [x] Disable job submission if active program has errors
- [x] Improve display/filtering of providers and only enable submission
on QIR capable ones.
- [x] Better code structure, error handling, and diagnostics logging.

TODO in a follow-up PR:

- Auto-refresh the workspace tree periodically if any jobs are in the
waiting or executing state, or newly submitted.
- Remember and automatically reload workspaces in the tree view across
editor sessions. (And ability to remove)
- Ability to filter, sort, and manage (cancel, delete, etc.) jobs.
- Ability to fetch the raw result data, display as histogram, etc.

---------

Co-authored-by: Mine Starks <mineyalc@microsoft.com>
Co-authored-by: Ian Davis <ian.f.davis@microsoft.com>
Co-authored-by: Sarah Marshall <33814365+samarsha@users.noreply.github.com>
Co-authored-by: Stefan J. Wernli <swernli@microsoft.com>